Update README.md to include hidden clean-css option
Hey, I noticed that when using this tool the clean-css step was removing IE specific CSS hacks which was causing display issues for me in IE. I've therefore switched off the cleanCss option, but noticed that the option wasn't mentioned in the docs. If there was valid reasoning for not putting this in the readme fair enough but thought it worth adding incase anyone else hits this issue.